requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
Pillow==9.3.0
tensorflow==2.10.0
moviepy==1.0.3
spacy==3.5.0
numpy==1.23.5
gunicorn==20.1.0
